Update human.py
Added the method get_choice. This method does not take any arguments and should return two values: the row and column (in order) to drop a bomb. Prompt the users for these values. The function re prompts the user if a bomb has already been dropped in the location selected.